backend/requirements.txt
fastapi>=0.115.0,<0.116.0
uvicorn[standard]>=0.30.0,<0.31.0
supabase>=2.4.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
